+++ docs/CHANGELOG.md
@@ -5,8 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.2.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.10.0]
+## [0.11.0]
 
+- Module changes:
+    - Meta
+        - Added support for Llama 3.1 405B Instruct - `meta.llama3-1-405b-instruct-v1:0`
+        - Updated pricing to reflect current Meta Llama 3.1 prices
+        - Corrected Meta Llama 3.1 models to show Multilingual support as `$true`
 - Build changes:
     - Updated tests to follow Pester 5 rules